+++ src/pages/formAccount.js
@@ -1,0 +1,10 @@
+import FormAccount from '../components/InputForms/FormAccount';
+
+export default function FormAccountPage() {
+	return (
+		<>
+			<h1>Your Bank Account</h1>
+			<FormAccount />
+		</>
+	);
+}
