--- src/pages/formBankAccount.js
@@ -1,10 +1,0 @@
-import AddNewBankAccount from '../components/AddNewAccount';
-
-export default function FormBankAccount() {
-	return (
-		<>
-			<h1>Your Bank Account</h1>
-			<AddNewBankAccount />
-		</>
-	);
-}
